pantheon-core/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyyaml==6.0.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
chromadb>=0.6.0
